Allow for no timestamp being appended
The documentation states that `false` is a valid value for the `dateFormat` global config param. That usually suggests that passing `false` means "do not add this field" but in this case, the code said otherwise, it meant: "use ISO".

Since most loggers nowadays (including `console.log`!) already add a timestamp to the log line, there has to be a way to opt-out of this field, to avoid repetition.

Example:

**Before:**
```
2020-08-11T21:25:35.736Z	90142447-1486-1b56-34e1-68e1b1439670	INFO	[Response] 2020-08-11T21:25:35+0000 https://example.com/api/meters/123 GET
```

**After:**
```
2020-08-11T21:25:35.736Z	90142447-1486-1b56-34e1-68e1b1439670	INFO	[Response] https://example.com/api/meters/123 GET
```